+++ lib/templates/bootstrap_vertical.ex
@@ -2,7 +2,11 @@ defmodule Formex.Template.BootstrapVertical do
   use Formex.Template, :main
   import Formex.Template.Bootstrap
 
-  def generate_row(form, field, template_options \\ []) do
+  @moduledoc """
+  The Bootstrap 3 [basic](http://getbootstrap.com/css/#forms-example) template. 
+  """
+
+  def generate_row(form, field, _options \\ []) do
 
     field_html  = generate_field_html(form, field)
     label_html  = generate_label_html(form, field)